Adding controller layer classes

+++ src/main/java/com/example/GuesthouseWebpage/service/BookingService.java
@@ -1,6 +1,7 @@
 package com.example.GuesthouseWebpage.service;
 
 
+import com.example.GuesthouseWebpage.exceptions.BookingNotFoundException;
 import com.example.GuesthouseWebpage.model.Booking;
 import com.example.GuesthouseWebpage.repository.BookingRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -25,22 +26,30 @@ public class BookingService {
         bookingRepository.save(booking);
     }
 
-
     //method to find booking by id
     public Optional<Booking> findBookingById(Long id) {
         return bookingRepository.findById(id);
     }
 
     //method to update booking by bookingId
-    public void updateBooking(Booking booking) {
+    public void updateBooking(Booking booking) throws BookingNotFoundException {
+        if (findBookingById(booking.getId()).isPresent()){
             bookingRepository.saveAndFlush(booking);
+        } else {
+            throw new BookingNotFoundException(booking.getId());
+        }
+
     }
 
     //method to "delete" booking by booking id - set active false so no one can see the booking
     public void deleteBookingById(Long id) {
             findBookingById(id).ifPresent(booking -> {
                 booking.setActive(false);
+                try {
                     updateBooking(booking);
+                } catch (BookingNotFoundException e) {
+                    e.printStackTrace();
+                }
             });
     }
 
@@ -48,13 +57,21 @@ public class BookingService {
     public void restoreBookingById(Long id){
         findBookingById(id).ifPresent(booking -> {
             booking.setActive(true);
-            updateBooking(booking);
+            try {
+                updateBooking(booking);
+            } catch (BookingNotFoundException e) {
+                e.printStackTrace();
+            }
         });
     }
 
     //method to really fully delete booking by booking id
-    public void fullDeleteBookingById(Long id) {
-        bookingRepository.deleteById(id);
+    public void fullDeleteBookingById(Long id) throws BookingNotFoundException {
+        if (findBookingById(id).isPresent()) {
+            bookingRepository.deleteById(id);
+        } else {
+            throw new BookingNotFoundException(id);
+        }
     }
 
     //list all bookings
@@ -68,6 +85,16 @@ public class BookingService {
                 .filter(Booking::isActive)
                 .collect(Collectors.toList());
     }
+
+    //list all confirmed bookings
+    public List<Booking> getConfirmedBookings(){
+        return getAllBookings().stream()
+                .filter(Booking::isBookingConfirmed)
+                .collect(Collectors.toList());
+    }
+
+
+
     //list all bookings in one date you receive from user
